Final_Bot_Binance/requirements.txt
python-binance
ccxt
pandas
numpy 
websocket_client